requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
cycler==0.11.0
Cython==0.29.30
flatbuffers==1.12
fonttools==4.33.3
gast==0.3.3
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
imageio==2.19.3
imgaug==0.4.0
importlib-metadata==4.11.4
joblib==1.1.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
keras-unet-collection==0.1.13
kiwisolver==1.4.3
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
networkx==2.8.4
numpy==1.23.0
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.8.1
Shapely==1.8.2
six==1.16.0
sklearn==0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tqdm==4.64.0
typing-extensions==3.7.4.3
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.12.1
zipp==3.8.0
